pyproject.toml
[project]
name = "WAZP"
authors = [
  {name = "Sofia Miñano", email = "s.minano@ucl.ac.uk"},
  {name = "Nikoloz Sirmpilatze", email = "niko.sirbiladze@gmail.com"},
  {name = "Sam Cunliffe", email = "s.cunliffe@ucl.ac.uk"},
  {name = "Adam Tyson", email = "code@adamltyson.com"}
]
description = "Wasp Animal-tracking Zoo project with Pose estimation"
readme = "README.md"
requires-python = ">=3.9.0"
dynamic = ["version"]

license = {text = "BSD-3-Clause"}

dependencies = [
  "tables",
  "blosc2",
  "numpy",
  "pillow",
  "pandas",
  "dash",
  "dash-bootstrap-components",
  "opencv-python",
  "PyYAML",
  "shapely",
  "openpyxl",
  "defusedxml",
  "pooch",
  "tqdm",
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: BSD License",
]

[project.urls]
"Homepage" = "https://github.com/SainsburyWellcomeCentre/WAZP"
"Bug tracker" = "https://github.com/SainsburyWellcomeCentre/WAZP/issues"
"Documentation" = "https://SainsburyWellcomeCentre.github.io/WAZP"
"Source code" = "https://github.com/SainsburyWellcomeCentre/WAZP"
"User support" = "https://github.com/SainsburyWellcomeCentre/WAZP/issues"

[project.scripts]
startwazp = "wazp.app:startwazp"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "coverage",
    "dash[testing]",
    "tox",
    "black",
    "ruff",
    "mypy",
    "pre-commit",
    "setuptools-scm",
    "types-PyYAML",
    "pandas-stubs",
    "types-Pillow",
]



[build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["wazp*"]
exclude = ["tests*", "docs*"]

[tool.pytest.ini_options]
addopts = "--cov=wazp"

[tool.black]
target-version = ['py39', 'py310']
skip-string-normalization = false
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | examples
  )/
)
'''

[tool.ruff]
line-length = 88
exclude = ["__init__.py","build",".eggs"]
select = ["I", "E", "F"]
fix = true

[tool.setuptools_scm]

[tool.check-manifest]
ignore = [
    "*.yaml",
    "tox.ini",
    "tests/*",
    "tests/test_unit/*",
    "tests/test_integration/*",
    "docs/",
    "docs/source/",
]

[[tool.mypy.overrides]]
module = [
  "dash.*",
  "dash_bootstrap_components.*",
  "h5py.*",
  "matplotlib.*",
  "plotly.*",
  "cv2.*",
  "shapely.*",
]
ignore_missing_imports = true
